Move not found page to [language] folder
