+++ site/src/app/[domain]/[language]/layout.tsx
@@ -1,5 +1,12 @@
+import { gql, previewParams } from "@comet/cms-site";
 import { IntlProvider } from "@src/app/[domain]/[language]/IntlProvider";
+import { GQLLayoutQuery, GQLLayoutQueryVariables } from "@src/app/[domain]/[language]/layout.generated";
+import { getSiteConfigForDomain } from "@src/config";
+import { Header } from "@src/layout/header/Header";
+import { headerFragment } from "@src/layout/header/Header.fragment";
+import { createGraphQLFetch } from "@src/util/graphQLClient";
 import { readFile } from "fs/promises";
+import { notFound } from "next/navigation";
 
 const messagesCache: Record<string, unknown> = {};
 async function loadMessages(language: string) {
@@ -10,10 +17,40 @@ async function loadMessages(language: string) {
     return messages;
 }
 
-export default async function Page({ children, params: { language } }: { children: React.ReactNode; params: { language: string } }) {
+export default async function Page({
+    children,
+    params: { domain, language },
+}: {
+    children: React.ReactNode;
+    params: { domain: string; language: string };
+}) {
+    const siteConfig = getSiteConfigForDomain(domain);
+
+    if (!siteConfig.languages.includes(language)) {
+        notFound();
+    }
+
     const messages = await loadMessages(language);
+
+    const { previewData } = (await previewParams()) || { previewData: undefined };
+    const graphqlFetch = createGraphQLFetch(previewData);
+
+    const { header } = await graphqlFetch<GQLLayoutQuery, GQLLayoutQueryVariables>(
+        gql`
+            query Layout($domain: String!, $language: String!) {
+                header: mainMenu(scope: { domain: $domain, language: $language }) {
+                    ...Header
+                }
+            }
+
+            ${headerFragment}
+        `,
+        { domain, language },
+    );
+
     return (
         <IntlProvider locale={language} messages={messages}>
+            <Header header={header} />
             {children}
         </IntlProvider>
     );
