+++ site/src/app/[domain]/[language]/not-found.tsx
@@ -1,4 +1,4 @@
-import { NotFoundContent } from "@src/app/not-found-content";
+import { NotFoundContent } from "@src/app/[domain]/[language]/not-found-content";
 import * as React from "react";
 
 export default function NotFound() {
